ARM: should use PRIu64 for uint64_t

+++ src/dhcp.c
@@ -2049,7 +2049,7 @@ int dhcp_dns(struct dhcp_conn_t *conn, uint8_t *pack,
 
 	snprintf(line, sizeof(line),
 		      "%d,"MAC_FMT",%s,%s,%d,%s\n",
-		      time(0),
+		      (int)time(0),
 		      MAC_ARG(conn->hismac),
 		      inet_ntoa(conn->hisip),
 		      q, authenticated,
